chore: refactor parse env
