+++ src/core/compare/parseAndFilterEnv.ts
@@ -1,4 +1,4 @@
-import { parseEnvFile } from '../parseEnv.js';
+import { parseEnvFile } from '../../services/parseEnvFile.js';
 import { filterIgnoredKeys } from '../filterIgnoredKeys.js';
 import type { ComparisonOptions } from '../../config/types.js';
