+++ src/index.ts
@@ -1,2 +1,3 @@
-export { parseEnvFile } from './core/parseEnv.js';
+export { parseEnvFile } from './services/parseEnvFile.js';
+export { parseEnvContent } from './core/parseEnv.js';
 export { diffEnv, type DiffResult } from './core/diffEnv.js';
