+++ test/unit/index.test.ts
@@ -1,11 +1,16 @@
 import { describe, it, expect } from 'vitest';
-import { diffEnv, parseEnvFile } from '../../src/index.js';
+import { diffEnv, parseEnvFile, parseEnvContent } from '../../src/index.js';
 import { diffEnv as diffEnvDirect } from '../../src/core/diffEnv.js';
-import { parseEnvFile as parseEnvFileDirect } from '../../src/core/parseEnv.js';
+import { parseEnvFile as parseEnvFileDirect } from '../../src/services/parseEnvFile.js';
+import { parseEnvContent as parseEnvContentDirect } from '../../src/core/parseEnv.js';
 
 describe('index exports', () => {
-  it('re-exports parseEnvFile from core/parseEnv', () => {
+  it('re-exports parseEnvFile from services/parseEnvFile', () => {
     expect(parseEnvFile).toBe(parseEnvFileDirect);
+  });
+
+  it('re-exports parseEnvContent from core/parseEnv', () => {
+    expect(parseEnvContent).toBe(parseEnvContentDirect);
   });
 
   it('re-exports diffEnv from core/diffEnv', () => {
